Multiple Changes From Empirical Study of Glycosaminoglycan data
While writing a small GAG deconvoluter, I found issues with a few parts of
the Link composition losses hard-coded to default to glycosidic losses
when the the passed losses were false-y. This behavior is no longer the
case, and the implementation only defaults to glycosidic losses if the
passed losses are explicitly `None`

I also improved the API for MolecularComposition, making them safe to
compare with strings. I also had to change the sigil character to `#` to
properly tokenize them.

Added empirical information for Anhydromannose substituents, which result
in a net loss of an Oxygen with respect to the canonical monosaccharide,
though I am not certain.